requirements.txt
Django==5.1.7
django-ninja==1.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
uvicorn==0.34.0
gunicorn==23.0.0
pandas==2.2.3
